build(deps): upgrade iced 0.13 -> 0.14
The one upgrade in this batch that needed real source changes. Grouped by
what 0.14 changed:

**`Pixels: From<u16>` removed.** This accounted for 187 of the 261 initial
errors, almost all of them flowing through the design tokens. `type_scale`,
`sidebar`, `spacing` and `shape` are logical-pixel values, so they become
`f32` and the project's own size/padding helpers (`icon`, `radius`,
`labeled`, `IconButton`, `Tag`, `TreeView::label_size`, `ActivityBadge`)
follow. `menu_panel_size` deliberately stays `u16`: it produces integer
window geometry for `clamp_menu_anchor`, which is pure, tested logic and
not a `Pixels` value — only its token reads are cast.

**`Widget` trait reshaped.** `layout`/`operate` take `&mut self` (so the
children go through `as_widget_mut`), `on_event` became `update` taking
`&Event` and returning `()`, and `overlay` gained a `viewport`. Captured
events are now signalled with `Shell::capture_event` instead of a returned
`event::Status`; every previous `Status::Captured` return is a
`capture_event()` + `return` pair, so the routing is unchanged.

**Widget/API renames.** `keyboard::on_key_press(f)` becomes
`keyboard::listen().filter_map(f)` — `filter_map` also requires a
zero-sized closure and hashes the closure's `TypeId`, so the
one-closure-per-overlay structure and its subscription identity both carry
over intact. Also `Space::new(w, h)` -> builder form, `checkbox(label,
checked)` -> `checkbox(checked).label(label)`, `progress_bar::height` ->
`girth`, canvas `Text::{horizontal,vertical}_alignment` -> `align_x`/
`align_y` (plus a new `max_width`), `window::get_latest`/`get_size` ->
`latest`/`size`, and `iced::application` now takes `boot` first with the
title as a builder and `run()` in place of `run_with(boot)`.

**Style structs grew fields.** `Palette::warning` (no warning role exists
in this UI, so it reuses `error`), `menu::Style::shadow` and
`scrollable`'s new `auto_scroll` overlay; the scroller thumb is a
`Background` rather than a `Color`. `pick_list::Status::Opened` and
`text_input::Status::Focused` are struct variants now. The zero shadow and
token-themed autoscroll keep the 0.13 look.

`stream::channel` takes an `AsyncFnOnce` in 0.14, so the daemon actor's
sender type has to be named rather than inferred.

Verified: `cargo fmt --check` clean, clippy clean across the workspace,
`mise run test` 781 passed / 0 failed (identical to the pre-upgrade
baseline), and the GUI runs for 60s under Wayland with no output. The
window's appearance was not visually diffed — no screenshot tool is
available on this machine.

+++ crates/micold-client/src/ui/material/icon_button.rs
@@ -21,9 +21,9 @@ type ButtonStyleFn = Box<dyn Fn(&iced::Theme, button::Status) -> button::Style>;
 pub struct IconButton<'a, M> {
     glyph: Icon,
     roles: Roles,
-    size: u16,
+    size: f32,
     tint: Option<Rgb>,
-    padding: u16,
+    padding: f32,
     circular: bool,
     on_press: Option<M>,
     _marker: PhantomData<&'a ()>,
@@ -46,14 +46,14 @@ impl<'a, M: Clone + 'a> IconButton<'a, M> {
     }
 
     /// Override the glyph size.
-    pub fn size(mut self, size: u16) -> Self {
+    pub fn size(mut self, size: f32) -> Self {
         self.size = size;
         self
     }
 
     /// Override the button's padding (defaults to `spacing::XS`) — widen the click target for a
     /// button that otherwise sits in a dense row (e.g. `spacing::SM`).
-    pub fn padding(mut self, padding: u16) -> Self {
+    pub fn padding(mut self, padding: f32) -> Self {
         self.padding = padding;
         self
     }
